Merged PR 178986: Config files required to onboard OneBranch Ownership Enforcer: repo.config and root directory owner.
Config files required to onboard OneBranch Ownership Enforcer: repo.config and root directory owner.
If we decided to enable this in cosmos-sdk, please enable it in VSTS UI.
EnableAadGroupOwner: Set this to true to allow security group as owner
DefaultHallpass: Set this to ture so that the reviewers who have already approved do not always have to re-approve new iterations again.