Merged PR 189340: Remove twitter_data.json from git history
Remove twitter_data.json from git history and and added .json files from TestJsons to git LFS. This rewrites all the SHA's from when that file was introduced. In order to merge this PR, I will have to do it instead of just hitting the complete button. Also anyone that as branches from master will have to do a rebase onto the new head which will be a new SHA and their head will be detached. The flag --allow-unrelated-histories will have to be used when doing a merge.

Related work items: #375521