Updating UpsertItemAsync
the original sample showed updating the value of AccountNumber (which is the partitionkey for this collection) on the retrieved item, and then doing an upsert

when i run this sample, it doesn't update the original item, it merely creates a new item, with a new partition key set but the old original item with the original partition key remains in the database. we effectively now have a duplicate document.
the observed behavior is as expected considering cosmosdb treats id + pk as unique for a document
therefore changing the pk value means cosmos db would never find the original item to update it, so it creates a new one.

if you run this sample, you end up with TWO SalesOrder3 items, 1 in the original partition, and another exactly the same but in a different partition

i believe the sample as it was is very misleading and confusing to users. you cannot change an item's partition this way, and if you do, you should go remove the old item that will still be there

this change modifies ShippedDate on the order, which results in the original item being updated, and not creating a duplicate which i believe is the intention of this sample.