Updating PR template with PR title as changelog
To reduce conflicts and changelog mistakes the PR titles will be used as the changelog in the future. This will make is easier for users looking through the history and avoid merge conflicts on the changelog file. This will require the changelog to get generated for each release, but can be done as part of contract validation.