[vNext] PR Requirements: Add PR title requirements for changelog generation (#1341)
* [Internal] PR Requirements: Add PR title requirements to allow changelog generation (#1322)

* Updating PR template with PR title as changelog

To reduce conflicts and changelog mistakes the PR titles will be used as the changelog in the future. This will make is easier for users looking through the history and avoid merge conflicts on the changelog file. This will require the changelog to get generated for each release, but can be done as part of contract validation.

* Updated based on feedback

* Fixed grammar

* More fixes

* Adding pr lint action

* Try single quotes to fix parse error

* Trying to move it to build yaml

* Try including all changes

* Updating

* Reverting changes

* updated based on feedback

* updating regex

* Removing space

* Updated based on user suggestion

* porting pr

* breaklines

* examples

* examples

Co-authored-by: j82w <j82w@users.noreply.github.com>