Private Preview AAD: Adds AAD support to the SDK (#1798)
Adds a constructor to take a TokenCredential

Adds a new TokenCredentialCache which store the AAD token while it is valid, and has a background task to refresh the token roughly 25% of the token lifespan.

AuthorizationTokenProvider is created as a base contract for all the authorization interfaces. Each authorization type now has it's own implementation. This avoids runtime exceptions and all the if else logic that is copy pasted everywhere.