Query: Fixes Equals method on SqlParameter class (#2044)
SqlParameterEquality relies on == operator on Name and Value. While this is okay for Name since that is explicitly typed as a string, it fails for value since that is typed as an object. For an object (even when it is a string) == falls to Object.ReferenceEquals, while the right thing to do is call the Equals method.