Query: Adds fast timeouts and retries to query plan (#2094)
PR #1954 only did retries on get operations. Get query plan is a post causing it to not do the faster retries. This adds explicit test to verify the behavior and to prevent future regression.

The diagnostics was updated to always include the retry policy to help trouble shoot issues in the future.