Internal PartitionKeyRangeCache: Adds RID refresh on NotFound in GetRoutingMapAsync (#2096)
In GetRoutingMapAsync, we read the RID from cache & then look up the associated collection routing map.
In the case where the RID is stale & the collection routing map is not found, the stale RID should be invalidated & the operation retried.

I am also switching a method written using ContinueWith to be async instead as I found during testing that this style can cause AggregateException to be thrown which callers do not typically expect (e.g. a NotFound would end up as a 500 because of that).