CosmosDiagnostics: Fixes IndexOutOfRangeException (#2099)
To avoid any IndexOutOfRangeException the CosmosDiagnostics is refactored to remove the AddRange call. This will avoid any indexing issues caused by the possibility of the CosmosDiagnosticsContext getting modified in another task at the same time it is getting append to a parent CosmosDiagnosticContext.

This should also help improve performance by reducing the number of copies required when combining multiple CosmosDiangosticsContext.