CosmosException: Fixes substatuscode to get the correct value instead of 0 when it is not in the enum. (#2110)
The document client exception that is thrown internally will return default value if the SubStatusCode is not defined in the enum. This changes the logic to parse and use the int from the headers so the exception always shows the correct value. This allows the SDK to properly handle unexpected or newer substatuscodes that it is not aware of.