Client Side Encryption : Adds Support for Managing Client Encryption Keys as a resource. (#2048)
This PR takes forward the Public Model of Client Side Data Encryption and is in conjunction with @anujtoshniwal PR [https://github.com//pull/2019] and focuses on managing Client Encryption Keys via Create/Replace as a resource and as part of a database and retrieve the same via ClientEncryptionKeyResponse as ClientEncryptionKeyProperties.

These methods are not meant to be invoked directly and will be used by Encryption Package to expose methods to manage ClientEncryptionKey. Please see https://aka.ms/CosmosClientEncryption in order to use client-side encryption.