CosmosClientOptions and CosmosClientBuilder: Adds client level support for EnableContentResponseOnWrite (#2145)
This PR adds a new field to the CosmosClient settings to enable/disable Content Response on write operations for Documents. Currently, this needs to be configured for every item request. We can now set this when creating CosmosClient. We can also override this in ItemRequestOptions while calling any Item operation.