[Preview] Item and Query Request Options: Adds Dedicated Gateway MaxIntegratedCacheStaleness (#2382)
Currently, the integrated cache (with the dedicated gateway) has a default expiration time for each of the entries in the cache (CacheEntryExpirationTime) of 5 minutes. Any response served from cache is guaranteed to be no staler than this CacheEntryExpirationTime. In the existing design, CacheEntryExpirationTime cannot be configured by customers.

For customers to have more control over how stale each of their requests could be, we plan to introduce a new “RequestOption” called “MaxIntegratedCacheStaleness”. This option will allow the customer to set a lower bound on the lifespan of the response from the cache for each request. If the lifetime of the response stored in the cache is greater than the value indicated by “MaxIntegratedCacheStaleness”, that item will not be served from the cache and the request will go to the backend.