Diagnostics: Adds Gateway Request Stats (#2324)
The current client side request stats does not record http operations. This is extending the contract to include all http operations which is necessary for troubleshooting latency issues. It was chosen to extend the IClientSideRequestStats rather than wiring though ITrace because of the dependencies in the direct package. Using ITrace would require wiring through all the shared code which is much larger task.