[Internal] Availability: Adds partition level retry logic for request timeout (#2446)
If a write request timeouts out it should retry on another region. It's possible that the current region is down and the backend has failed over to another region. This will ensure that the SDK will failover in such scenarios.

Write request can not be retried. It's not clear if the request made it to the service or not.
Only being used for write requests as if it is not the primary region it will give a fast and clear write forbidden exception. This allows the SDK to figure out the correct primary region.
Improved the trace messages to include the location and resource address. This is needed to help which SDK instance is hitting the client retry policy.