Availability: Fixes account refresh logic on gateway outage (#2447)
The previous behavior would not refresh the account information if there was a gateway outage, and it would depend on the background refresh to update the information. The background refresh can take up to 5 minutes before detecting a change.

HttpRequestExceptions will trigger an account refresh every 15 seconds. A config was added to override the minimum time between refreshes.
Account refresh will be a no-op if a different thread is already updating the information
Updated the traces to include the name to make it easier to discover
GlobalEndpointManager is refactored to use 2 methods instead of different overloads on 1 method. This makes it clear that 1 method loads the account information and starts the background refresh. The other is forcing a refresh outside the background refresh.