[Internal] Engineering: Refactors code to reduce technical debt (#2470)
This pull request is very simple and aims to reduce the technical debt. It specifically targets the "Duplicate casts should not be made" rule in SonarQube.

Update:

As pointed out by @ealsur, the SonarQube fixes touched OSS files that probably should't be modified. To make this contribution more valuable, I fixed the following ReSharper issues:

"Replace if statement with null-propagating code" rule
"Merge sequential checks in && or || expressions" rule