Availability: Fixes get account info retry logic to not go to secondary regions on 401(Forbidden) (#2511)
This fixes a bug where if a user gets a 401(Forbidden) exception the SDK will try to go to the secondary regions. It will also now throw an aggregate exception if there are multiple failures instead of just throwing the last. This needed because it's possible the last failure maybe different then the cause of the other failures.