[Internal] HDR histogram: Removes unit tests (#2504)
The HDR Histogram tests are flaky causing the gates to randomly fail. There has so far been 5 failures on 5 different tests on 5 different runs. It is also increasing the time for the gates to run. This PR is removing the unit tests. They are not needed because the OSS clone should not be modified.