CosmosDiagnostics: Fixes a race condition causing InvalidOperationException (#2516)
This is a regression caused by making the ClientSideRequestStatisticsTraceDatum private fields to public. Previously these lists were protected by locks and when they were made public the locks were removed. This PR is adding the locks back and adding tests to ensure that if a background thread does update it while the list is being enumerated that it will not throw a InvalidOperationCancelledException.

Having a lock in the IEnumerable adds the possibility of deadlocks if the enumerator is not properly disposed of. To avoid this possibility a shallow copy of the list is lazily created. Once it is generated it will reuse the same shallow copy until the list is modified again.