CosmosException: Adds diagnostics to message for 408, 500, 503, 404/1002 scenarios (#2719)
This change is being done because most users are not capturing the diagnostics for these exceptions. It is not possible to root cause or trouble shoot these issues without the diagnostics. This requires the customer to make code changes to get the additional info which causes the issue to persist for a long time.