Availability: Adds EnableTcpConnectionEndpointRediscovery to true by default (#2724)
EnableTcpConnectionEndpointRediscovery is used to enable address cache refresh on TCP connection reset notification. By refreshing the caches it helps future requests from going to the stale addresses.

ConnectionStateListener holds a cache that maps the server physical addresses to the partitionKeyRanges(Server Key -> List[PartitionKeyRangeIdentity]). ReplicatedResourceClient updates this cache when it resolves the physical addresses for the partitions. When the listener receives a notification on connection reset from the transport client, it triggers an address refresh in the GatewayAddressCache for all the partitionKeyRanges that maps to the server address received in the notification.