Query: Fixes a bug where max page size is not being honored after the first 2 pages (#2776)
Query results were not honouring MaxItemCount after 2 pages. This causing the query to be slow and cost a lot of RUs. 
As part of this PR fixing the bug.

## Analysis:
It was found that we are loosing configured MaxItemCount somewhere in the flow and getting set as Int32 Max i.e. 2,147,483,647 which was leading to above issue. 
So in this _OrderByQueryPartitionRangePageAsyncEnumerator_, We have _InnerEnumerator_ which stores _QueryPaginationOptions_ and there we have private field and one public field to store the same information. In the constructor of this, we were setting only private field and getter was getting value from the public field. So changed the code to set the public variable and remove the private variable.