Availability: Adds nonblocking async lazy cache to improve upgrade scenarios (#2822)
The current AsyncCache on force refreshes removes the value. It then creates a new value and all incoming requests are stuck waiting for the new task to complete.

Issue
If a single replica moves only 1 of the 4 replica URIs are stale. The current async cache is blocking all requests until the value is updated. The 3 other replicas could be used to successfully complete requests.

Solution
AsyncCacheNonBlocking was created. This async cache does not block requests on refresh. It instead returns the stale value until the refresh is complete. Then the refresh value is used to update the cache. Any request using the forceRefresh to true will wait on the same refresh task.

Porting from internal PR: https://msdata.visualstudio.com/CosmosDB/_git/CosmosDB/pullrequest/648930