[Internal] Diagnostics: Adds test for Point Operation With Service Unavailable Exception (#2833)
The following diagnostics is from the 3.18.1 SDK. This hit an exception which seems to have cause the diagnostics to be improperly nested. The "Microsoft.Azure.Documents.ServerStoreModel Transport Request" should be at the same level but are some how nested. This needs to be root caused and fixed if it hasn't already been fixed in the latest SDK version.

Added Point Operation With Service Unavailable Exception and verified the output json in test method. Now its generating ServerStoreModel once . There was no nesting issue found.