Benchmark: Adds the config to the run summary (#2912)
The run summary before was only including a subset of the configs. This makes it difficult to determine if a regression was caused by a change in the config. The run summary will now include the full config which ignores secret fields. This way if new configs get added they will automatically get logged with each run.