Benchmark: Adds loop script and update run.sh (#2918)
This PR adds a loop.sh which is used to update to the latest git commit and execute the run.sh.

The run.sh script is updated to add scenarios for client telemetry and to make query operations optional. This is necessary because query runs take too long to be included in every run.