[Internal] Client Telemetry : Adds ThreadPool Information (#2923)
We already get Thread pool information in our diagnostics. So As part of this PR, this information will be aggregated and pushed as part of client telemetry.

Metric Name Convention: < Thread Pool Name >_< Type of Metric > As Java has multiple threadpools so "ThreadPool name" will be useful there. But for .Net we decided to keep it a generic name i.e SystemPool.