Initialization: Fixes the SDK to retry if the initialization fails (#3027)
The SDK has an initialization task to get the account information and other info. If this task fails with a DocumentClientException it was never being recreated. This is a problem because if a 408 is thrown or some other transient problem the SDK will always returned the cached failure and will not actually retry the request.

Solution:
A new initialization function factory was created. This allows the initialization task to be recreated. The EnsureValidClientAsync now calls a new method that is thread safe to get or create a new task if the existing one failed.