[INTERNAL]  Line endings: Adds gitattribute to normalize line endings (#3062)
Several of the files have mixed line endings causing issues with PRs saying they changed the entire file instead of just 1 line. This is adding a gitattribute file that will force all the files to use the CRLF formatting. This also fixes all the .cs files to use the correct formatting.