Availability: Adds retry on 408s to gateway calls for metadata operations (#3089)
This applies the same retry logic for a 408 from gateway that is already applied if the a timeout occurs trying to connect to the gateway. This will improve availability to customers because the SDK will not automatically retry on a timeout from gateway for metadata operation. This will only be used for cache operations where multiple requests are waiting on the result and a single timeout would result in multiple failures.