[Internal] LocalQuorum : Adds configuration to allow strong read with eventual consistency account (#3071)
SDK doesn't allow overriding with stronger consistency config. So it means if Account level consistency is Set to "Eventual" then you can not request for a request/client with "Strong" consistency. But vice versa is allowed.
Since Back-end supports it now, it should pass this particular validation in SDK also.

### Solution
Introduce a new flag `EnableUpgradeConsistencyToLocalQuorum` ( `AllowUpgradeConsistencyToLocalQuorum` function in Builder), if this flag is true then, it will pass the validation for such requests.