Upsert item: Fixes a bug causing session tokens for PKRangeId 0 to be overridden by session token for another partition on upserts (#3102)
TransportHandler.ProcessUpsertAsync was capturing session tokens even for throttled (or failed) requests. Because throttled requests (or 449/410) won't always have a PKRangeId header this resulted in session tokens for PKRangeId 0 being overridden with the session token of another partition. There is no reason to ever capture a session token on a 429, 449 or 410 - and it turned out there was also no need for the special-casing of upserts in TransportHandler.

So, this PR is...

removing the special casing of upserts in TransportHandler
adding regression test (NoSessionTokenCaptureForThrottledUpsertRequestsTest failed without the change in TransportHandler - now succeeds)
adding basic coverage to ensure session tokens are captured for point operations (main validation for this is in Direct package)