Add Key Updates for CosmosClient
Adds the ability to update the account key in a `CosmosClient` via a `CosmosMasterKeyCredential`. The implementation follows model similar to that used in Azure.Storage's [`StorageSharedKeyCredential`](https://github.com/Azure/azure-sdk-for-net/blob/main/sdk/storage/Azure.Storage.Common/src/StorageSharedKeyCredential.cs) class.

A client constructed using a `CosmosMasterKeyCredential` will utilize that credential to generate the token hashes. The `CosmosMasterKeyCredential.UpdateKey` method allows the key (and the resulting hash function) to be updated so all subsequent requests utilize the new hash function.

The underlying implementation of `AuthorizationTokenProviderMasterKey` has been updated to use a `CosmosMasterKeyCredential` instead of keeping its own `IComputeHash` function. Note that if the `CosmosClient` was constructed using a `CosmosMasterKeyCredential` then the credential is *not* disposed when the client is disposed. However, if the client was constructed using an auth key directly, then the resulting `CosmosMasterKeyCredential` created internally will be disposed with the client since it is owned by the client.