Diagnostics: Adds response serialization time (#3165)
Add trace to track total serialization time in responses for additional debugging info.

Fixes a bug where if a start child is called on a disposed trace which would then cause the timing to be off. It will now walk up the trace tree until it finds the active parent and add the child to that parent to keep the structure valid.