LINQ: Fixes NullReferenceException when attempting to generate SQL for a LINQ expression which performs static field or property access (fixes #3188) (#3190)
This PR should fix the issue raised in #3188 in which attempts to access a static field or property (such as DateTime.UtcNow) within a LINQ-to-SQL expression will result in a NullReferenceException. It does so by correctly handling the presence of a null member expression when evaluating field and property access.