[Preview] AI Integration: Adds ability to emit open telemetry attributes (#3189)
This is the first PR as part of a series to add open telemetry support to the preview SDK. This PR adds the overloads to allow client context to log the open telemetry. Future PRs will wire E2E when all scenarios are supported. This will not log anything to open telemetry currently. 

Enable SDK to emit open telemetry attributes on different operations. This PR will cover all the basic implementation of Open Telemetry. under Preview tag.