Update templates/build-test.yml
I am committing your suggestion. Also, the goal of this PR is not to executed long running tests. I will create a new PR for this. Because there are other unknowns that I need to answer. So for now, I just don't want it to be executed with the existing pipelines.

Co-authored-by: Matias Quaranta <ealsur@users.noreply.github.com>